+++ mrv2/docs/HISTORY.md
@@ -2,9 +2,12 @@ v0.3.7
 ------
 - Added a half OCIO default to handle OpenEXR half images.
 - Added timeRange and inOutRange to timeline module.
-- Added setIn() and setOut() to timeline module to set the in and out time/frame/seconds.
+- Added setIn() and setOut() to timeline module to set the in and out
+  time/frame/seconds.
 - Fixed timeRange conversion to string (__str__) and repr (__repr__).
 - Fixed Presentation toggle from the menus and from the right mouse button menu.
+- Added media.firstVersion(), media.previousVersion(), media.nextVersion(),
+  and media.lastVersion() to move from one version of the clip to the next.
 
 v0.3.6
 ------